requirements.txt
pyannote.audio
sqlalchemy
openai-whisper
click
ncls
imagehash
pillow
opencv-python-headless
scipy
tqdm